+++ mail/mail-uniffi/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<<<<<<< HEAD
 ## [mail-uniffi-v0.162.6] - 2026-02-20
 
 ### Fixes
@@ -21,12 +22,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [mail-uniffi-v0.162.4] - 2026-02-16
+## [mail-uniffi-v0.161.28] - 2026-02-27
+
+### Fixes
+
+- Tie scroller order with local_id for very rare instance of returning duplicate entries
+
+
+## [mail-uniffi-v0.161.27] - 2026-02-24
+>>>>>>> releases/mail-uniffi/0.161
 
 ### Fixes
 
 - [ET-5882] Missing handler for event provider fetches
 
 
+<<<<<<< HEAD
 ## [mail-uniffi-v0.162.3] - 2026-02-10
 
 ### Fixes
@@ -34,6 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Conflicts between Cargo.lock
 
 
+=======
+>>>>>>> releases/mail-uniffi/0.161
 ## [mail-uniffi-v0.161.26] - 2026-02-10
 
 ### Fixes